backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
fastapi==0.115.8
h11==0.14.0
idna==3.10
numpy==2.2.3
pandas==2.2.3
psycopg2==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
typing_extensions==4.12.2
tzdata==2025.1
uvicorn==0.34.0
